Refactor the filters
Previously the filters were applies when entering a directory. To apply
a filter to the current directory, one would leave it and reenter with
the new filter. It was done so because the code responsible for focusing
the first visible file was in the deer-enter function. Now the code was
extracted from it and the filter changing is independent of entering to
the directory.

As a bonus, it is now possible to apply a filter to the root directory.